kops: migrate all kernels to HF kernel-builder (AOT, nix, Hub-ready)
All 9 custom kernels are now kernel-builder repos under src/kops/<name>/
(build.toml + csrc + torch-ext with a native TORCH_LIBRARY op + flake.nix/lock),
replacing the load_inline JIT path. src/kops/registry/<name>.py are thin
get_local_kernel loaders (Python prep for nms/gaussian_blur/moe/deform/roi stays
there). Built via nix on the GB10 Spark (scripts/build_kernels.sh); all 9 op-bench
verify and beat/tie op_compile. build/ is gitignored + transfer-excluded.

Docs/skill rewritten for the kernel-builder flow; ruff/pyrefly exclude the generated
torch-ext packages. Hub repo-ids are placeholders (kernels-local/*) — not pushed;
see HANDOFF.md for push-readiness.

+++ src/kops/deform_attn/csrc/deform_attn.cu
@@ -1,4 +1,4 @@
-#include <torch/extension.h>
+#include <torch/all.h>
 #include <ATen/cuda/CUDAContext.h>
 
 // Fused multi-scale deformable attention forward — performance-tuned.
@@ -106,7 +106,7 @@ __global__ void ms_deform_attn_fwd_kernel(const float *__restrict__ value, const
 }
 
 at::Tensor ms_deform_attn_forward(at::Tensor value, at::Tensor spatial_shapes, at::Tensor level_start_index,
-                                  at::Tensor sampling_locations, at::Tensor attention_weights, int im2col_step) {
+                                  at::Tensor sampling_locations, at::Tensor attention_weights, int64_t im2col_step) {
     int B = value.size(0);
     int L_total = value.size(1);
     int n_heads = value.size(2);
